Migrate repository to more descriptive naming
