+++ packaging/debroot/DEBIAN/conffiles
@@ -9,4 +9,4 @@
 /etc/rundeck/rundeck-config.properties
 /etc/rundeck/ssl/ssl.properties
 /etc/rundeck/apitoken.aclpolicy
-/etc/rundeck/verb-repositories.yaml
+/etc/rundeck/artifact-repositories.yaml
