+++ rundeckapp/grails-spa/src/pages/repository/main.js
@@ -1,0 +1,18 @@
+// The Vue build version to load with the `import` command
+// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
+import Vue from 'vue'
+import Vue2Filters from 'vue2-filters'
+import VueCookies from 'vue-cookies'
+import App from './App'
+
+Vue.config.productionTip = false
+
+Vue.use(Vue2Filters)
+Vue.use(VueCookies)
+
+/* eslint-disable no-new */
+new Vue({
+  el: '#repository-vue',
+  components: { App },
+  template: '<App/>'
+})
