--- rundeckapp/grails-spa/src/pages/verb/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>Verb Artifact Browser</title>
-    <link rel="stylesheet" href="http://localhost:4440/assets/bootstrap.min.css?compile=false">
-    <link rel="stylesheet" href="http://localhost:4440/assets/app.scss.css?compile=false">
-    <script type="text/javascript">
-      window._rundeck = {
-        rdBase: 'http://localhost:4440/',
-        offset: '0',
-        limit: '20'
-      }
-    </script>
-  </head>
-  <body>
-    <div id="verb-vue"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
