+++ rundeckapp/repository/grails-app/conf/plugin.yml
@@ -1,0 +1,8 @@
+---
+rundeck:
+    features:
+        repository:
+            enabled: false
+    repository:
+        plugins:
+        artifacts:
