+++ rundeckapp/repository/grails-app/controllers/repository/ArtifactController.groovy
@@ -1,0 +1,6 @@
+package repository
+
+class ArtifactController {
+
+    def index() { }
+}
