+++ rundeckapp/repository/grails-app/i18n/messages.properties
@@ -1,0 +1,1 @@
+gui.menu.FindPlugins=Find Plugins
