+++ rundeckapp/repository/grails-app/init/repository/BootStrap.groovy
@@ -1,0 +1,12 @@
+package repository
+
+class BootStrap {
+
+    def grailsApplication
+
+    def init = { servletContext ->
+        log.debug("Repository enabled: " + grailsApplication.config.rundeck.features.repository.enabled)
+    }
+    def destroy = {
+    }
+}
