--- rundeckapp/verb/grails-app/conf/plugin.yml
@@ -1,5 +1,0 @@
----
-rundeck:
-    features:
-        verb:
-            enabled: false
