--- rundeckapp/verb/grails-app/controllers/verb/VerbController.groovy
@@ -1,6 +1,0 @@
-package verb
-
-class VerbController {
-
-    def index() { }
-}
