--- rundeckapp/verb/grails-app/i18n/messages.properties
@@ -1,1 +1,0 @@
-gui.menu.Verb=Verb Plugins
