--- rundeckapp/verb/grails-app/init/verb/Application.groovy
@@ -1,12 +1,0 @@
-package verb
-
-import grails.boot.*
-import grails.boot.config.GrailsAutoConfiguration
-import grails.plugins.metadata.*
-
-@PluginSource
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
