--- rundeckapp/verb/grails-app/init/verb/BootStrap.groovy
@@ -1,12 +1,0 @@
-package verb
-
-class BootStrap {
-
-    def grailsApplication
-
-    def init = { servletContext ->
-        log.debug("Verb enabled: " + grailsApplication.config.rundeck.features.verb.enabled)
-    }
-    def destroy = {
-    }
-}
